pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"
requires-python = ">=3.10,<3.11"

[project]
name = "hackaton"
version = "0.1.0"
description = "Hackaton Gen AI Project"
dependencies = [
    "openai",
    "python-dotenv",
    "langchain",
    "langchain-openai",
    "langchain-core",
    "langchain-community",
    "gradio"
]
